split a ts file
